CSRF prevention
This upgrades to Spring Security 3.2 RC1. Yes, it's a release candidate,
but the API is pretty stable, and it's the only way to get native CSRF
token support.

Note that X-Frame-Options is set to DENY by default. If you're using
OpenEssence in a frame, you'll have to do some (small) tweaks.